refactor: ポケモン管理システムをシンプル版に大幅簡素化
## 変更内容
### ER図とデータベース設計の簡素化
- テーブル数を8個から4個に削減
- 複雑な属性を最小限に絞り込み
- 技システム、ボックスシステム、個体値システムを削除

### データベーススキーマ
- pokemon_master: 基本ステータス（HP、攻撃、防御）のみ
- owned_pokemon: レベルと現在HPのみ管理
- party_pokemon: パーティ管理のみ
- 複雑な制約やトリガーを削除

### 型定義とリポジトリ
- simple-pokemon.tsで簡素化された型定義を作成
- simplePokemonRepositoryで基本CRUD操作のみ実装
- シンプルなステータス計算ロジック
- TDD approach維持

### 設計方針
- 最低限動作するポケモンライクゲームを目標
- 初学者が理解しやすいシンプルな構造
- 必要に応じて後から機能拡張可能な設計
